fix(playerbot): Fix namespace issues, covariant returns, and stub TODO methods
- Fixed GroupRole enum member (MELEE_DPS -> DPS_MELEE)
- Fixed syntax error in GroupCoordinator.cpp (missing if statement)
- Fixed malformed includes in QuestStrategy.cpp and LootStrategy.cpp
- Added namespace Playerbot declaration to ArenaAI.h
- Fixed GroupCoordinator namespace syntax
- Added GameTime.h include to GroupCoordinator.cpp
- Fixed TacticalCoordinator namespace (Playerbot::Advanced::)
- Fixed IntegratedAIContext return types for TacticalCoordinator
- Moved ArenaMetrics to namespace scope to avoid covariant return type conflict
- Fixed QuestCompletionMetrics return type inconsistency
- Added GroupEvent forward declaration and include
- Stubbed out incompatible GroupCoordinator methods (GetBotsByRole, IsInCombat, GetCombatDuration, GetFocusTarget)
- Stubbed out RaidOrchestrator GroupCoordinator creation (per-bot architecture incompatible)
- Fixed PvPCombatAI.h include placement (moved outside namespace)
