+++ src/modules/Playerbot/AI/Integration/IntegratedAIContext.cpp
@@ -63,7 +63,7 @@ SharedBlackboard* IntegratedAIContext::GetZoneBlackboard() const
     return _cachedZoneBlackboard;
 }
 
-TacticalCoordinator* IntegratedAIContext::GetTacticalCoordinator() const
+Advanced::TacticalCoordinator* IntegratedAIContext::GetTacticalCoordinator() const
 {
     if (!_cachedTacticalCoordinator && _bot)
     {
@@ -229,7 +229,7 @@ BTStatus BTCheckGroupFocusTarget::TickWithContext(IntegratedAIContext& context)
 
 BTStatus BTFollowGroupStrategy::TickWithContext(IntegratedAIContext& context)
 {
-    TacticalCoordinator* coordinator = context.GetTacticalCoordinator();
+    Advanced::TacticalCoordinator* coordinator = context.GetTacticalCoordinator();
     if (!coordinator)
         return BTStatus::FAILURE;
